requirements.txt
Django==3.0
djangorestframework==3.13.1